moved recipeDetail to views folder and renamed the file
