+++ src/views/RecipeDetailView.vue
